Ch4V14

+++ Java Advanced Project 4/SpeedConverter/src/com/company/Main.java
@@ -3,7 +3,7 @@ package com.company;
 public class Main {
     public static void main(String[] args) {
 
-        long miles = SpeedConverter.toMilesPerHour(  10.5);
+        double miles = SpeedConverter.toMilesPerHour(  10.5);
         System.out.println("Miles = " +miles);
 
         SpeedConverter.printConversion(  10.5);
